utils: rename to util
